Conflicts:
	ASMHelper
	java/squeek/applecore/asm/module/ModuleBlockFood.java
	java/squeek/applecore/asm/module/ModuleFoodEatingSpeed.java
	java/squeek/applecore/asm/module/ModuleFoodStats.java
	java/squeek/applecore/asm/module/ModulePlantGrowth.java

+++ java/squeek/applecore/asm/reference/ItemRendererModifications.java
@@ -22,8 +22,8 @@ public class ItemRendererModifications
 		// this block is deep in the method
 		if (enumaction == EnumAction.eat || enumaction == EnumAction.drink)
 		{
-			f6 = (float) entityclientplayermp.getItemInUseCount() - p_78440_1_ + 1.0F;
-			f7 = 1.0F - f6 / (float) entityclientplayermp.itemInUseMaxDuration;
+			f6 = entityclientplayermp.getItemInUseCount() - p_78440_1_ + 1.0F;
+			f7 = 1.0F - f6 / entityclientplayermp.itemInUseMaxDuration;
 
 			// ...
 		}
